Performance and reliability improvements
If split vector cannot be used to read a collection, fall back to a
paginated splits. Add mode AnyOrder for when the order coming off the
oplog is not important.  This allows for sending inserts and deletes
immediately off the oplog instead of waiting for whole document to be
fetched from updates.  Flush the update queue more often.