Reduce number of connections generated
This commit puts a limit on the number of splits the split vector
command with return and matches this number in the paginating
splitter. Since the number of connections to MongoDB will be splits+1
for each collection read from, we don't want the number of splits to get
too high.  Also this re-introduces query batching with was recently
removed.